+++ config/environments/production.rb
@@ -47,7 +47,7 @@ Rails.application.configure do
   config.active_support.report_deprecations = false
 
   # Replace the default in-process memory cache store with a durable alternative.
-  case ENV.fetch("RAILS_CACHE_STORE", Settings.cache&.store)
+  case ENV.fetch("RAILS_DATABESE_IN_MEMORY", Settings.database&.in_memory || "solid")
   in "solid"
     config.cache_store = :solid_cache_store
   in "redis"
@@ -55,13 +55,13 @@ Rails.application.configure do
       url: ENV.fetch("REDIS_URL", "redis://localhost:6379/0"),
       namespace: "ikamail:cache",
     }
-  else
-    Rails.logger.debug("Using default cache store: :memory_store")
-    # config.cache_store = :memory_store
+  in "memory"
+    Rails.logger.debug("Using volatile cache store: :memory_store")
+    config.cache_store = :memory_store
   end
 
   # Replace the default in-process and non-durable queuing backend for Active Job.
-  case ENV.fetch("RAILS_QUEUE_ADAPTER", Settings.queue&.adapter)
+  case ENV.fetch("RAILS_DATABESE_IN_MEMORY", Settings.database&.in_memory || "solid")
   in "solid"
     config.active_job.queue_adapter = :solid_queue
     # solid_queue configuration
@@ -69,11 +69,12 @@ Rails.application.configure do
     # mission_control-jobs configuration
     config.mission_control.jobs.base_controller_class = "AdminController"
     config.mission_control.jobs.http_basic_auth_enabled = false
-  in "resque"
+  in "redis"
+    # use resque with redis backend
     config.active_job.queue_adapter = :resque
-  else
-    Rails.logger.debug("Using default Active Job queue adapter: :async")
-    # config.active_job.queue_adapter = :async
+  in "memory"
+    Rails.logger.debug("Using volatile Active Job queue adapter: :async")
+    config.active_job.queue_adapter = :async
   end
 
   # Ignore bad email addresses and do not raise email delivery errors.